pyproject.toml
[tool.poetry]
name = "hackatons"
version = "0.1.0"
description = ""
authors = ["kirill <shkolenko02-03-2004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-cors-headers = "4.3.0"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
drf-spectacular = {extras = ["sidecar"], version = "^0.26.5"}
gunicorn = "^21.2.0"
djangorestframework-simplejwt = "^4.7.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
psycopg = {extras = ["binary"], version = "^3.1.13"}
django-filter = "^23.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
